+++ test-parts/test-pathfinding-input.js
@@ -6,10 +6,13 @@ const parseGraph = require("../src/parse-graph");
 const exactRoute = require("../src/exact-route");
 const shortestRoute = require("../src/shortest-route");
 const possibleRoutes = require("../src/possible-routes");
+const numSigns = require("../src/common/num-signs");
+const routeCriteria = require("../src/common/route-criteria");
 
 const graphErrMsg = "Cannot read property 'nodes' of null";
 const unknownNodesMsg = "UNKNOWN NODES";
 const noRouteMsg = "NO SUCH ROUTE";
+const badCriteriaMsg = "INVALID COUNT CRITERIA";
 
 var exampleGraph = null;
 
@@ -19,8 +22,9 @@ function runTests()
 	describe("Invalid Pathfinding Input", function()
 	{
 		prepareExampleGraph();
-		handleExactRoute();
-		handleShortestRoute();
+		//handleExactRoute();
+		//handleShortestRoute();
+		handlePossibleRoutes();
 	});
 }
 
@@ -149,6 +153,81 @@ function handleShortestRoute()
 }
 
 
+function handlePossibleRoutes()
+{
+	describe("Possible Routes", function()
+	{
+		var emptyCriteria = null;
+		var invalidCriteria = routeCriteria.defineCriteria("Bad number", "Bad sign");
+		
+		it("Correct Output - Single", function()
+		{
+			var stopInput = routeCriteria.defineCriteria(1, numSigns.EQUAL);
+			var distInput = routeCriteria.defineCriteria(10, numSigns.LESS_EQUAL);
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "C", stopInput, distInput);
+			expect(resultValue).to.equal(1);
+		});
+		
+		it("Correct Output - Multiple", function()
+		{
+			var stopInput = routeCriteria.defineCriteria(5, numSigns.LESS_EQUAL);
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "F", stopInput, emptyCriteria);
+			checkMultiplePossibleRoutes(resultValue);
+		});
+		
+		it("Correct Output - Zero", function()
+		{
+			var distInput = routeCriteria.defineCriteria(5, numSigns.LESS_EQUAL);
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "F", emptyCriteria, distInput);
+			expect(resultValue).to.equal(0);
+		});
+		
+		it("Missing Graph", function()
+		{
+			callPossibleRoutesMissingGraph();
+		});
+		
+		it("Missing Node Argument", function()
+		{
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "D", null, emptyCriteria, emptyCriteria);
+			expect(resultValue).to.equal(unknownNodesMsg);
+		});
+		
+		it("Empty Node Argument", function()
+		{
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "", null, emptyCriteria, emptyCriteria);
+			expect(resultValue).to.equal(unknownNodesMsg);
+		});
+		
+		it("Unknown Node", function()
+		{
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "O", "P", emptyCriteria, emptyCriteria);
+			expect(resultValue).to.equal(unknownNodesMsg);
+		});
+		
+		
+		it("Invalid Distance Criteria", function()
+		{
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "B", emptyCriteria, invalidCriteria);
+			expect(resultValue).to.equal("INVALID DISTANCE CRITERIA");
+		});
+		
+		it("Invalid Stop Count Criteria", function()
+		{
+			var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "B", invalidCriteria, emptyCriteria);
+			expect(resultValue).to.equal(badCriteriaMsg);
+		});
+		
+		it("Impossible Route", function()
+		{
+			//var resultValue = possibleRoutes.findRoutes(exampleGraph, "A", "G", emptyCriteria, emptyCriteria);
+			//console.log(resultValue);
+			expect(true).to.be.true;
+		});
+		
+	});
+}
+
 
 function callExactRouteMissingGraph()
 {
@@ -175,6 +254,25 @@ function callShortestRouteMissingGraph()
 	try
 	{
 		shortestRoute.findRoute(null, "A", "B");
+		givenMessage = "";
+	}
+	catch(routeErr)
+	{
+		givenMessage = routeErr.message;
+	}
+	
+	checkMissingGraphError(givenMessage);
+}
+
+
+function callPossibleRoutesMissingGraph()
+{
+	var givenMessage = "";
+	
+	try
+	{
+		possibleRoutes.findRoutes(null, "A", "B", null, null);
+		givenMessage = "";
 	}
 	catch(routeErr)
 	{
@@ -196,9 +294,23 @@ function checkOutputDistanceNumber(distNum)
 	}
 	else
 	{
-		throw Error("Output must be a positive, whole number.");
+		throw new Error("Output must be a positive, whole number.");
 	}
+}
+
+
+function checkMultiplePossibleRoutes(countNum)
+{
+	var correctType = Number.isInteger(countNum);
 	
+	if (correctType === true && countNum >= 2)
+	{
+		expect(true).to.be.true;
+	}
+	else
+	{
+		throw new Error("Output must be a whole number that is at least 2.");
+	}
 }